use mkdocs-gen-files over custom hooks
