--- docs/.gitignore
@@ -1,5 +1,0 @@
-# Generated by hooks/cli.py
-references/cli.md
-
-# Generated by hooks/rest_api.py
-references/openapi.json
